Refactor code structure; Remove unused files and functions; set default model path
