--- src/identifiers/__init__.py
@@ -1,1 +1,0 @@
-"""Identifiers used to identify various document types."""
