+++ src/models/treebased/train.py
@@ -1,4 +1,5 @@
 import argparse
+import json
 import logging
 import os
 import time
@@ -10,10 +11,10 @@ from sklearn.model_selection import RandomizedSearchCV
 from swissgeol_doc_processing.utils.file_utils import read_params as swissgeol_read_params
 from xgboost import XGBClassifier
 
-from classifiers.pdf_dataset_builder import build_filename_to_label_map
 from models.treebased.basetrainer import TreeBasedTrainer
 from models.treebased.model_explanation import explain_model
 from src.models.feature_engineering import get_features
+from src.page_classes import label2id
 from src.utils.utility import get_pdf_files, read_params
 
 logger = logging.getLogger(__name__)
@@ -105,6 +106,26 @@ def load_data_and_labels(folder_path: Path, label_map: dict[tuple[str, int], int
                 labels.append(label_map[key])
 
     return all_features, labels
+
+
+def build_filename_to_label_map(gt_json_path: Path) -> dict[tuple[str, int], int]:
+    """Build a map from filename to class ID based on the ground truth JSON."""
+    with open(gt_json_path) as f:
+        gt_data = json.load(f)
+
+    label_lookup = {}
+    for entry in gt_data:
+        filename = entry["filename"]
+        for pages in entry["pages"]:
+            page = pages["page"]
+            for label_name, value in pages["classification"].items():
+                if value == 1:
+                    try:
+                        label_id = label2id[label_name]
+                        label_lookup[(filename, page)] = label_id
+                    except KeyError as err:
+                        raise ValueError(f"Unknown label: {label_name}") from err
+    return label_lookup
 
 
 def main(config_path: str, out_directory: str, tuning: bool = False):
